Move helper routines to utils
