+++ src/xtbml/utils/utils.py
@@ -28,3 +28,19 @@ def timing(f):
         return result
 
     return wrap
+
+
+@torch.jit.script
+def real_atoms(numbers: Tensor) -> Tensor:
+    return numbers != 0
+
+
+@torch.jit.script
+def real_pairs(numbers: Tensor, diagonal: bool = False) -> Tensor:
+    real = real_atoms(numbers)
+    mask = real.unsqueeze(-2) * real.unsqueeze(-1)
+    if not diagonal:
+        mask *= ~torch.diag_embed(torch.ones_like(real))
+    return mask
+
+
